refactor: Home.jxs to Home.tsx
